requirements.txt
PyPDF2
python-dotenv
streamlit
openai
altair
tiktoken
protobuf
ninja
torch
sentence-transformers
transformers
langchain-huggingface
llama-index-embeddings-huggingface
llama-index-embeddings-instructor
llama-index-embeddings-openai
llama-index
llama-index-embeddings-ollama
llama-index-vector-stores-chroma
xformers